# Conflicts:
#	doc/user_guide/index.rst
#	hyperspy/axes.py
#	hyperspy/drawing/image.py
#	hyperspy/tests/axes/test_data_axis.py
#	hyperspy/tests/model/test_model.py

+++ hyperspy/_signals/signal2d.py
@@ -507,7 +507,7 @@ class Signal2D(BaseSignal, CommonSignal2D):
                 min(images_number, chunk_size)
             pcarray = ma.zeros((nrows, self.axes_manager._max_index + 1,
                                 ),
-                               dtype=np.dtype([('max_value', np.float),
+                               dtype=np.dtype([('max_value', float),
                                                ('shift', np.int32,
                                                 (2,))]))
             nshift, max_value = estimate_image_shift(
